+++ Sources/Onboarding/Screens/Welcome/Subviews/BottomSection.swift
@@ -12,7 +12,6 @@ struct BottomSection<C: View> {
     private let accentColor: Color
     private let appDisplayName: String
     private let continueAction: () -> Void
-    private let signInWithAppleConfiguration: SignInWithAppleButtonConfiguration?
     private let dataPrivacyContent: () -> C
     @State private var isDataPrivacyPresented: Bool = false
     @State private var isAnimating: Bool = false
@@ -21,13 +20,11 @@ struct BottomSection<C: View> {
         accentColor: Color,
         appDisplayName: String,
         continueAction: @escaping () -> Void,
-        signInWithAppleConfiguration: SignInWithAppleButtonConfiguration? = nil,
         @ViewBuilder dataPrivacyContent: @escaping () -> C
     ) {
         self.accentColor = accentColor
         self.appDisplayName = appDisplayName
         self.continueAction = continueAction
-        self.signInWithAppleConfiguration = signInWithAppleConfiguration
         self.dataPrivacyContent = dataPrivacyContent
     }
 
@@ -94,20 +91,13 @@ extension BottomSection: View {
 
     @ViewBuilder
     private var continueButton: some View {
-        if let signInWithAppleConfiguration {
-            SignInWithAppleButtonView(
-                configuration: signInWithAppleConfiguration,
-                continueAction: continueAction
-            )
-        } else {
-            Button(
-                action: continueAction,
-                label: continueText
-            )
-            .font(.title3.weight(.medium))
-            .buttonStyle(.borderedProminent)
-            .tint(accentColor)
-        }
+        Button(
+            action: continueAction,
+            label: continueText
+        )
+        .font(.title3.weight(.medium))
+        .buttonStyle(.borderedProminent)
+        .tint(accentColor)
     }
 
     private func continueText() -> some View {